Refactoring the CLI stuff into it's own package
