+++ autorest/adal/persist.go
@@ -1,14 +1,11 @@
 package adal
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-
-	"github.com/dimchansky/utfbom"
 )
 
 // LoadToken restores a Token object from a file located at 'path'.
@@ -26,42 +23,6 @@ func LoadToken(path string) (*Token, error) {
 		return nil, fmt.Errorf("failed to decode contents of file (%s) into Token representation: %v", path, err)
 	}
 	return &token, nil
-}
-
-// LoadCLITokens restores a set of AzureCLIToken objects from a file located at 'path'.
-func LoadCLITokens(path string) ([]AzureCLIToken, error) {
-	file, err := os.Open(path)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open file (%s) while loading token: %v", path, err)
-	}
-	defer file.Close()
-
-	var tokens []AzureCLIToken
-
-	dec := json.NewDecoder(file)
-	if err = dec.Decode(&tokens); err != nil {
-		return nil, fmt.Errorf("failed to decode contents of file (%s) into a AzureCLIToken representation: %v", path, err)
-	}
-
-	return tokens, nil
-}
-
-// LoadCLIProfile restores an AzureCLIProfile object from a file located at 'path'.
-func LoadCLIProfile(path string) (AzureCLIProfile, error) {
-	var profile AzureCLIProfile
-
-	contents, err := ioutil.ReadFile(path)
-	if err != nil {
-		return profile, fmt.Errorf("failed to open file (%s) while loading token: %v", path, err)
-	}
-	reader := utfbom.SkipOnly(bytes.NewReader(contents))
-
-	dec := json.NewDecoder(reader)
-	if err = dec.Decode(&profile); err != nil {
-		return profile, fmt.Errorf("failed to decode contents of file (%s) into a AzureCLIProfile representation: %v", path, err)
-	}
-
-	return profile, nil
 }
 
 // SaveToken persists an oauth token at the given location on disk.
